Extract abstract base class to be used for Entry implementations
